requirements.txt
streamlit==1.30.0
pandas==2.1.0
transformers==4.35.0
torch==2.1.0
sqlalchemy==2.0.23
pandas==2.1.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
psycopg2-binary==2.9.7
mysql-connector-python==8.2.0
pydantic==2.5.0
loguru==0.7.2
uvicorn==0.24.0
fastapi==0.104.0